challenge_2/pyproject.toml
[tool.poetry]
name = "challenge-2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "~3.10"
jupyter = "^1.0.0"
numpy = "^1.23.5"
scikit-learn = "^1.2.0"
tensorflow-macos = "2.10"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"
flake8 = "^6.0.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
